AI-ML-project/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.9
numpy>=1.26.0
pandas>=2.1.0
requests>=2.31.0
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
pytest>=7.4.0
flake8>=6.1.0
black>=23.9.1
tensorflow>=2.14.0
